Updates
1) Preload first 10 comments for post
2) Fix error messages for user info editing
3) Remove download button for personal data
4) Add blocked user list in settings
5) Restrict blocked user from writing comments and reacting
6) Add email change
7) Fix general bugs

+++ src/MicroBlog.Data.EF/Repositories/PostRepository.cs
@@ -18,8 +18,13 @@ namespace MicroBlog.Data.EF.Repositories
         public override async Task DeleteRange(IEnumerable<Post> models)
         {
             var ids = models.Select(x => x.Id).ToHashSet();
-            await DbSet.Where(x => ids.Contains(x.Id))
-                .ExecuteDeleteAsync().ConfigureAwait(false);
+            await DbSet.Where(x => ids.Contains(x.Id)).ExecuteDeleteAsync().ConfigureAwait(false);
+            await Db.SaveChangesAsync().ConfigureAwait(false);
+        }
+
+        public async Task DeleteForUser(string userId)
+        {
+            await DbSet.Where(x => x.UserId == userId).ExecuteDeleteAsync().ConfigureAwait(false);
             await Db.SaveChangesAsync().ConfigureAwait(false);
         }
     }
